refactored the VC generation into a separate file
